docs: add note about config schema instability
Document that we consider the configuration file schema to be unstable instead of supporting versioning or strict backwards compatibility. Thanks to the strict schema validation, necessary configuration changes upon updating `talos-bootstrap` will be (relatively) clearly highlighted to the user, and we do not need to take on the maintenance burden of legacy format support. Since the user chooses when to update, schema evolution should not cause a constant burden to them either.

Signed-off-by: Dennis Marttinen <twelho@welho.tech>