feat: support configuring Gateway API
Adds support for configuring the native [Gateway API support](https://docs.cilium.io/en/stable/network/servicemesh/gateway-api/gateway-api/) in Cilium. Configuration options are provided for binding directly to node ports instead of relying on a LoadBalancer, as well as privilege configuration for enabling binding to privileged ports. `talos-bootstrap` will NOT automatically install the Gateway API CRDs in your cluster. Cilium requires this to be done separately, refer to the documentation for details.

Signed-off-by: Dennis Marttinen <twelho@welho.tech>